create multiple-choice headers to replace symlink mechanism
the build system used to symlink different target-specific directories
into the source tree so the would be picked up.
this is quite a mess and requires that configure does the job, if it's
done from the Makefile we can't tell whether the links are correctly
set up.

following script was used to create the headers:

import sys

pfx = sys.argv[1]
header = sys.argv[2]

for i in range(3, len(sys.argv)):
        if i == 3: word = "if"
        else: word = "elif"
        a = sys.argv[i]
        d = a.split('-', 1)[1]
        s = '#%s defined(%s_%s)\n#include "%s/%s"\n'%(word, pfx, d.replace('-', '_').upper(), a, header)
        sys.stdout.write(s)

example invocation:

    python foo.py MACHINE machdep.h md-68k md-i386-gcc md-ppc md-amd64-gcc md-ppc-gcc md-x86-gcc md-generic