revamp build system
- replace automake with hand-written Makefile
- strip down configure.ac

the previous build system was a maintenance nightmare:
recursive make, symlinked-in-place target directories, dozens of
configuration options and like 6**6 different backend configuration
possibilities.

caveat emptor: it's well possible that we'll have to go back
several times to inspect the previous configure.ac for specific
preprocessor macros enabling/disabling a certain code path.

at the moment the build is known working in the following configuration:

./configure --with-threads=posix --with-sound=alsa --with-machine=amd64-gcc

the md-generic machine backend is currently broken, after starting UAE
it'll just wait forever for vblank to happen.

it's planned to further cleanup the configure script, but removing each
of the checks involves cleaning up all the users of the provided autoconf
cpp macro, for example HAVE_STDIO_H.