import od-win32 directory from WinUAE v 2.8.1
as of commit 615d9dcb (release 2.8.1).

the purpose of having these files (even though we don't use them)
is so we can apply a complete diff from upstream - often changes
in the emulator core are accompanied by related changes in the
GUI, and without seeing those in the commit history it's often
hard to reason what a specific change is trying to achieve.

the price to pay is 8.6 MB of uncompressed sources, probably around
2-3 MB in raw .git data.