gtk-gui: temporary hack to make it possible to switch rom
the whole range of "start" "stop" "pause" buttons are currently broken,
as the emulator just starts in parallel with the GTK gui and there
seems to be no way implemented to communicate these running state
between GUI and engine (for example UAECMD_START, GUICMD_PAUSE etc
are only ever used in the GTK ui itself).
the proper fix would be, just like WinUAE does, to not start the
emulator window before the START button is hit in the GUI, and
make the GUI disappear until it is invoked with F12.
however this is harder as it sounds as the startup code for the
gui and the emulator core are heavily interwoven.
the proper place to fix is probably real_main2 in main.c.

until that is fixed, we just ignore the emulator state and make it
always possible to select another rom, so the user can load rom and
floppy and hit "reset" (the only emulator state button that actually works)
to make the emulator work.

(the intended way seems to be to have a rom called "kick.rom" in roms/, or
to start the emulator with command line switches or a config file)